+++ flight/targets/pipxtreme/board-info/pios_board.h
@@ -254,7 +254,7 @@ extern uintptr_t pios_com_frsky_sport_id;
 //-------------------------
 
 #if defined(PIOS_INCLUDE_RFM22B)
-extern uint32_t pios_rfm22b_id;
+extern uintptr_t pios_rfm22b_id;
 #endif /* PIOS_INCLUDE_RFM22B */
 
 //-------------------------
